# backend/openapi.json
{"openapi":"3.0.0","paths":{"/":{"get":{"operationId":"ping","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"string"}}}}}}},"/ping/llm":{"get":{"operationId":"pingLLM","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object"}}}}}}},"/ping/gen-api":{"get":{"operationId":"pingGenApi","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object"}}}}}}},"/stories":{"post":{"operationId":"create","summary":"Create a new story","parameters":[],"responses":{"201":{"description":"The story has been successfully created.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/StoryDto"}}}}}}},"/stories/{id}":{"get":{"operationId":"findOne","summary":"Get a story by ID","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"The story has been successfully retrieved.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/StoryDto"}}}}}}},"/stories/{id}/generate-content":{"post":{"operationId":"generateChapterContent","summary":"Generate content for a story","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"The content has been successfully generated.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/StoryDto"}}}},"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/StoryDto"}}}}}}},"/stories/{id}/generate-chapter-media":{"post":{"operationId":"generateStoryChapterMedia","summary":"Generate chapter media for a story","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"The media has been successfully generated.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/StoryDto"}}}},"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/StoryDto"}}}}}}},"/stories/{id}/generate-media":{"post":{"operationId":"generateFullStoryMedia","summary":"Generate media for a story","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"The media has been successfully generated.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/StoryDto"}}}},"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/StoryDto"}}}}}}},"/stories/{id}/generate-background-image":{"post":{"operationId":"generateStoryBackgroundImage","summary":"Generate background image for a story","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"The background image has been successfully generated.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/StoryDto"}}}},"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/StoryDto"}}}}}}},"/types":{"get":{"operationId":"findAllTypes","parameters":[],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TypeDto"}}}}}}},"/types/{id}":{"get":{"operationId":"findOneType","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}}}},"/types/{id}/stories":{"get":{"operationId":"findAllStoriesByType","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"type":"object"}}}}}}}},"/publishing/{storyId}":{"get":{"operationId":"getPublishing","parameters":[{"name":"storyId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PublishingDto"}}}}}}},"/publishing/{storyId}/youtube":{"post":{"operationId":"publishYoutube","parameters":[{"name":"storyId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"201":{"description":"","content":{"application/json":{"schema":{"type":"object"}}}}}}},"/youtube/auth":{"get":{"operationId":"getYoutubeAuthUrl","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"string"}}}}}}},"/youtube/logout":{"get":{"operationId":"youtubeLogout","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object"}}}}}}},"/youtube/auth/callback":{"get":{"operationId":"youtubeAuthCallback","parameters":[{"name":"code","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object"}}}}}}}},"info":{"title":"Story Generator","description":"The story generator API description","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"schemas":{"ChapterDto":{"type":"object","properties":{"id":{"type":"number"},"number":{"type":"number"},"title":{"type":"string"},"summary":{"type":"string"},"content":{"type":"string"},"background_sound":{"type":"string"},"audio_url":{"type":"string"},"video_url":{"type":"string"},"stories_id":{"type":"number"}},"required":["id","number","title","summary","content","background_sound","audio_url","video_url","stories_id"]},"PublishingDto":{"type":"object","properties":{"id":{"type":"number"},"story_id":{"type":"number"},"title":{"type":"string"},"description":{"type":"string"},"tags":{"type":"array","items":{"type":"string"}},"youtube_id":{"type":"string"},"patreon_published":{"type":"boolean"}},"required":["id","story_id","title","description","tags","youtube_id","patreon_published"]},"TypeDto":{"type":"object","properties":{"id":{"type":"number"},"name":{"type":"string","description":"Type of the story","example":"Horror"},"story_prompt":{"type":"string"},"chapter_prompt":{"type":"string"},"image_prompt":{"type":"string"},"sound_prompt":{"type":"string"},"youtube_channel_id":{"type":"string"},"youtube_playlist_id":{"type":"string"}},"required":["id","name","story_prompt","chapter_prompt","image_prompt","sound_prompt","youtube_channel_id","youtube_playlist_id"]},"StoryDto":{"type":"object","properties":{"id":{"type":"number","description":"ID of the story","example":1},"name":{"type":"string","description":"Title of the story","example":"The Haunting of Hill House"},"synopsis":{"type":"string","description":"Synopsis of the story","example":"A family moves into a haunted house and discovers its dark secrets."},"audio_url":{"type":"string"},"image_prompt":{"type":"string"},"background_image":{"type":"string"},"thumbnail_url":{"type":"string"},"video_url":{"type":"string"},"chapters":{"description":"Array of chapters in the story","type":"array","items":{"$ref":"#/components/schemas/ChapterDto"}},"publishings":{"type":"array","items":{"$ref":"#/components/schemas/PublishingDto"}},"types_id":{"type":"number","description":"ID of the story type","example":1},"types":{"$ref":"#/components/schemas/TypeDto"},"created_at":{"format":"date-time","type":"string"},"updated_at":{"format":"date-time","type":"string"}},"required":["name","synopsis","audio_url","image_prompt","background_image","thumbnail_url","video_url","chapters","types_id","types"]}}}}